% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_fun.R
\name{IsoPlot_fun}
\alias{IsoPlot_fun}
\title{IsoPlot function}
\usage{
IsoPlot_fun(
  feat_table = NULL,
  metadata_table = NULL,
  groups = "sample",
  relativeCID = TRUE,
  dodgeCID = FALSE,
  outpath = "./isoplot_results",
  pdf = TRUE,
  img = TRUE
)
}
\arguments{
\item{feat_table}{Features table}

\item{metadata_table}{Metadata table}

\item{groups}{Grouping factor}

\item{relativeCID}{Stacked barplot of CID with relative value (TRUE) or absolute value (FALSE)}

\item{dodgeCID}{Stacked barplot of CID with dodge position (TRUE) or stack position (FALSE)}

\item{outpath}{Output path}

\item{pdf}{Output PDF (TRUE) or not (FALSE)}

\item{img}{Output JPG (TRUE) or not (FALSE)}
}
\value{
A list containing the following elements:

\itemize{

\item \strong{Isoplot} A list of ggplot objects containing the barplots of the experimental isotopologue fraction of each metabolite with error bars
\item \strong{EnrC13_TotArea} A list of ggplot objects containing the barplots of the mean EnrC13 and Total Area of each metabolite (one figure per metabolite)
\item \strong{EnrC13_TotArea_allMet} A list of ggplot objects containing the barplots of the mean EnrC13 and Total Area of all metabolites (one figure per sample or group of samples)
}
}
\description{
A function to plot isotopic data from metabolomic experiments
}
\examples{
\dontrun{
feat <- glue::glue(system.file(package = "graphstatsr"), "/dataset/isoplot_quantification_table.csv")
metadata <- glue::glue(system.file(package = "graphstatsr"), "/dataset/isoplot_metadata.csv")
res <- IsoPlot_fun(feat_table = feat, metadata_table = metadata, groups = "sample", 
	relativeCID = TRUE, outpath = "./isoplot_results")

}

}
