% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPT_fun.R
\name{theoretical_abundances}
\alias{theoretical_abundances}
\title{Theoretical abundance calculation function}
\usage{
theoretical_abundances(n, k, p)
}
\arguments{
\item{n}{Number of carbons in the molecule}

\item{k}{Number of the isotopologue}

\item{p}{Abundance of labeled precursor}
}
\value{
Theoretical abundance of isotopologues
}
\description{
Calcuate the theoretical abundance of isotopologues
}
\keyword{#'}
\keyword{0.513);}
\keyword{0:10,}
\keyword{<-}
\keyword{@examples}
\keyword{internal}
\keyword{plot(res)}
\keyword{res}
\keyword{theoretical_abundances(10,}
