% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_func.R
\name{TP_func}
\alias{TP_func}
\title{Theoretical abundance calculation function}
\arguments{
\item{path}{Path to the input file (TSV format)}

\item{p}{Natural abundance of 13C (default = 0.513)}

\item{outpath}{Path to the output directory (default = "./MSPT_out/"), if NULL, no output is generated}
}
\value{
A list containing the following elements:

\itemize{

\item \strong{Table} A data frame containing the results of the calculations (sample, metabolite, isotopologue, area, Total_Area, Ratio, Theoretical_Ratios, Mean_Ratios, Mean_Ratios_SD, Thresholds, Bias (%), Mean Bias (%), Mean Bias SD (%))
\item \strong{figures} A list of ggplot objects containing the barplots of the experimental and theoretical isotopologue fraction of each metabolite with error bars
}
}
\description{
This function calculates the theoretical abundance of isotopologues
}
\examples{
\dontrun{
file <- glue::glue(system.file(package = "graphstatsr"), "/dataset/MSPT_test.tsv")
res <- TP_func(path = file, outpath = NULL)
str(res, max.level = 2)
}

}
