% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPT_fun.R
\name{MSPT_fun}
\alias{MSPT_fun}
\title{MSPT function}
\usage{
MSPT_fun(
  path,
  p = 0.513,
  outpath = "./MSPT_out/",
  minCID = 0.02,
  maxBias = 5,
  feature = NULL
)
}
\arguments{
\item{path}{Path to the input file (TSV format)}

\item{p}{Natural abundance of 13C (default = 0.513)}

\item{outpath}{Path to the output directory (default = "./MSPT_out/"), if NULL, no output is generated}

\item{minCID}{Minimum value of CID to be considered (default = 0.02)}

\item{maxBias}{Maximum value of bias to be considered (default = 5)}

\item{feature}{Name of the metabolite to preview (default = NULL)}

\item{plotLowCID}{Logical, if TRUE, plot the isotopologues with CID < minCID (default = FALSE) (not implemented yet)}
}
\value{
A list containing the following elements:

\itemize{

\item \strong{Table} A data frame containing the results of the calculations (sample, metabolite, isotopologue, area, Total_Area, Ratio, Theoretical_Ratios, Mean_Ratios, Mean_Ratios_SD, Thresholds, Bias (%), Mean Bias (%), Mean Bias SD (%))
\item \strong{figures} A list of ggplot objects containing the barplots of the experimental and theoretical isotopologue fraction of each metabolite with error bars
}
}
\description{
This function calculates the isotopologue fraction of each metabolite and compares it with the theoretical isotopologue fraction. It generates a table with the results and barplots of the experimental and theoretical isotopologue fraction of each metabolite with error bars.
}
\examples{
\dontrun{
file <- glue::glue(system.file(package = "graphstatsr"), "/dataset/MSPT_test.tsv")
res <- MSPT_fun(path = file, outpath = NULL)
str(res, max.level = 2)
}

}
